Moved new nondet slice functions to a separate file
